requirements.txt
requests
bs4 
datetime 
python-dotenv
numpy
boto3
pandas
praw
textblob




